InfoCenter Base Functionality
InfoCenter should now supports getTasks, which returns an ArrayList
containing all Tasks, and addTask, which adds a single task to the
table.

InfoCenter.initialize() MUST be called before any other methods.

NOTE: Need to add a text file "database.properties" with the required
property information - the url, the username, and the password.  I'm a
little shaky on what exactly is required for that, kinda need to talk
through it, or have another example to look at.

Also, I have no good way to test this at the moment, so there might be a
bug or two.