InfoCenter Update/Remove functions & Task ID
Added updateTask and removeTask to InfoCenter, which updates or removes
a provided task from the table respectively.

Also added an ID field and get/set methods to the Task class for table
location identification. It has NOT been added to the constructors, so
only the InfoCenter class should have anything to do with it.

Also, a few syntax errors I noticed have been rectified.